+++ src/Hyperar.HPA.Database/senior/Tables/MatchEvent.sql
@@ -1,18 +1,20 @@
 ﻿CREATE TABLE [senior].[MatchEvent] (
-    [Id]                         INT     IDENTITY (1, 1) NOT NULL,
-    [Index]                      TINYINT  NOT NULL,
-    [PlayerHattrickId]           BIGINT  NULL,
-    [AdditionalPlayerHattrickId] BIGINT  NULL,
-    [TeamHattrickId]             BIGINT  NULL,
-    [Type]                       SMALLINT  NOT NULL,
-    [Variation]                  SMALLINT  NOT NULL,
-    [Text]                       NTEXT   NULL,
-    [Minute]                     TINYINT  NOT NULL,
-    [MatchPart]                  TINYINT NOT NULL,
-    [MatchHattrickId]            BIGINT  NOT NULL,
+    [Id]                         INT    IDENTITY (1, 1) NOT NULL,
+    [Index]                      INT    NOT NULL,
+    [PlayerHattrickId]           BIGINT NULL,
+    [AdditionalPlayerHattrickId] BIGINT NULL,
+    [TeamHattrickId]             BIGINT NULL,
+    [Type]                       INT    NOT NULL,
+    [Variation]                  INT    NOT NULL,
+    [Text]                       NTEXT  NULL,
+    [Minute]                     INT    NOT NULL,
+    [MatchPart]                  INT    NOT NULL,
+    [MatchHattrickId]            BIGINT NOT NULL,
     CONSTRAINT [PK_MatchEvent] PRIMARY KEY CLUSTERED ([Id] ASC),
     CONSTRAINT [FK_Senior_MatchEvent_Match] FOREIGN KEY ([MatchHattrickId]) REFERENCES [senior].[Match] ([HattrickId])
 );
+
+
 
 
 GO
