+++ src/Hyperar.HPA.Database/senior/Tables/MatchTeamBooking.sql
@@ -1,15 +1,17 @@
 ﻿CREATE TABLE [senior].[MatchTeamBooking] (
     [Id]               INT            IDENTITY (1, 1) NOT NULL,
-    [Index]            TINYINT        NOT NULL,
+    [Index]            INT            NOT NULL,
     [PlayerHattrickId] BIGINT         NOT NULL,
     [PlayerName]       NVARCHAR (256) NOT NULL,
-    [Type]             TINYINT        NOT NULL,
-    [Minute]           TINYINT        NOT NULL,
-    [MatchPart]        TINYINT        NOT NULL,
-    [MatchTeamId]           INT            NOT NULL,
+    [Type]             INT            NOT NULL,
+    [Minute]           INT            NOT NULL,
+    [MatchPart]        INT            NOT NULL,
+    [MatchTeamId]      INT            NOT NULL,
     CONSTRAINT [PK_MatchTeamBooking] PRIMARY KEY CLUSTERED ([Id] ASC),
     CONSTRAINT [FK_Senior_MatchTeamBooking_MatchTeam] FOREIGN KEY ([MatchTeamId]) REFERENCES [senior].[MatchTeam] ([Id])
 );
+
+
 
 
 GO
