+++ src/Hyperar.HPA.Database/senior/Tables/PlayerSkillSet.sql
@@ -1,22 +1,24 @@
 ﻿CREATE TABLE [senior].[PlayerSkillSet] (
-    [Id]               INT      IDENTITY (1, 1) NOT NULL,
-    [Season]           SMALLINT NOT NULL,
-    [Week]             TINYINT  NOT NULL,
-    [Form]             TINYINT  NOT NULL,
-    [Stamina]          TINYINT  NOT NULL,
-    [Keeper]           TINYINT  NOT NULL,
-    [Defending]        TINYINT  NOT NULL,
-    [Playmaking]       TINYINT  NOT NULL,
-    [Winger]           TINYINT  NOT NULL,
-    [Passing]          TINYINT  NOT NULL,
-    [Scoring]          TINYINT  NOT NULL,
-    [SetPieces]        TINYINT  NOT NULL,
-    [Loyalty]          TINYINT  NOT NULL,
-    [Experience]       TINYINT  NOT NULL,
-    [PlayerHattrickId] BIGINT   NOT NULL,
+    [Id]               INT    IDENTITY (1, 1) NOT NULL,
+    [Season]           INT    NOT NULL,
+    [Week]             INT    NOT NULL,
+    [Form]             INT    NOT NULL,
+    [Stamina]          INT    NOT NULL,
+    [Keeper]           INT    NOT NULL,
+    [Defending]        INT    NOT NULL,
+    [Playmaking]       INT    NOT NULL,
+    [Winger]           INT    NOT NULL,
+    [Passing]          INT    NOT NULL,
+    [Scoring]          INT    NOT NULL,
+    [SetPieces]        INT    NOT NULL,
+    [Loyalty]          INT    NOT NULL,
+    [Experience]       INT    NOT NULL,
+    [PlayerHattrickId] BIGINT NOT NULL,
     CONSTRAINT [PK_PlayerSkillSet] PRIMARY KEY CLUSTERED ([Id] ASC),
     CONSTRAINT [FK_PlayerSkillSet_Player] FOREIGN KEY ([PlayerHattrickId]) REFERENCES [senior].[Player] ([HattrickId])
 );
+
+
